+++ CMakeLists.txt
@@ -38,7 +38,7 @@ SET(CMAKE_CXX_STANDARD 14)
 SET(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 IF (BUILD_COVERAGE)
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fprofile-arcs -ftest-coverage")
+  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fprofile-arcs -ftest-coverage -DBUILD_COVERAGE")
 ENDIF()
 
 # Generate documentation
@@ -65,7 +65,7 @@ ADD_SUBDIRECTORY(ext)
 
 INCLUDE_DIRECTORIES(${YAML_CPP_INCLUDE_DIR})
 
-SET(GALS_INCLUDE "${CMAKE_CURRENT_SOURCE_DIR}/src/include")
+SET(GALS_INCLUDE "${CMAKE_CURRENT_SOURCE_DIR}/include")
 INCLUDE_DIRECTORIES(${GALS_INCLUDE})
 
 ADD_LIBRARY(gals_cpu SHARED ${GALS_CPU_SRC})
