+++ docs/CMakeLists.txt
@@ -7,6 +7,7 @@ SET(DOCS_LAYOUT "${CMAKE_CURRENT_SOURCE_DIR}/layout.xml")
 SET(DOCS_LAYOUT_OUT "${CMAKE_CURRENT_BINARY_DIR}/layout.xml.out")
 
 SET(DOCS_DIR     "${CMAKE_CURRENT_SOURCE_DIR}")
+SET(INCLUDE_DIR "${PROJECT_SOURCE_DIR}/include")
 SET(SRC_DIR "${PROJECT_SOURCE_DIR}/src")
 #set(EXAMPLES_DIR "${PROJECT_SOURCE_DIR}/examples")
 SET(SNIPPETS_DIR "${PROJECT_SOURCE_DIR}/tests")
