+++ include/gals/cpu/gradient.h
@@ -31,6 +31,8 @@
 
 #pragma once
 
+#include "gals/cpu/gradient/second-order-central.h"
+#include "gals/cpu/gradient/third-order.h"
 #include "gals/utilities/array.h"
 #include "gals/utilities/grid.h"
 #include "gals/utilities/mat3.h"
@@ -40,23 +42,23 @@ namespace GALS
 {
 namespace CPU
 {
-/*! \class ThirdOrder
+/*! \class Gradient
  *
- * Gradient computation using third order scheme.
+ * Class to create 3 component elements at a computational cell. For e.x. velocity, gradients, etc.
  */
-template <typename T, typename T_GRID>
-class ThirdOrder
+template <typename T, typename T_GRID, typename GRADIENT_SCHEME = ThirdOrder<T, T_GRID>>
+class Gradient
 {
  public:
-  typedef T value_type;
+  using value_type = T;
 
   /*! Default constructor
    */
-  ThirdOrder();
+  Gradient();
 
   /*! Destructor
    */
-  ~ThirdOrder();
+  ~Gradient();
 
   /*! Compute gradient of scalar.
    *
@@ -66,44 +68,19 @@ class ThirdOrder
    * \param alpha scalar variable for which gradient will be computed.
    * \param grad_alpha gradient of alpha is written to this.
    */
-  void compute(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha);
+  static void compute(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha);
 
-  /*! Overload operator to compute gradient of a scalar quantity.
-   *
-   * For a scalar the gradient is stored in Vec3, for e.x.:
-   *  \f$\left(\nabla \alpha\right)_i = \{\partial_x \alpha, \partial_y \alpha, \partial_z \alpha\}_i.\f$
-   *
-   * \param alpha scalar variable for which gradient will be computed.
-   * \param grad_alpha gradient of alpha is written to this.
-   */
-  void operator()(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha) { compute(alpha, grad_alpha); }
-
-  /*! Compute gradient of Vec3.
+  /*! Compute gradient of vector.
    *
    * For a Vec3 the gradient is stored in Mat3, for e.x.:
    *  Let \f$\boldsymbol{u} = \{u, v, w\} \f$ then,
    *  \f$\left(\nabla \boldsymbol{u}\right)_i = \begin{bmatrix}\partial_x u & \partial_y u & \partial_z u \\ \partial_x
    * v & \partial_y v & \partial_z v \\ \partial_x w & \partial_y w & \partial_z w \end{bmatrix}_i.\f$
    *
-   * \param alpha Vec3 variable for which gradient will be computed.
+   * \param alpha vec3 variable for which gradient will be computed.
    * \param grad_alpha gradient of alpha is written to this.
    */
-  void compute(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha);
-
-  /*! Overload operator to compute gradient of Vec3.
-   *
-   * For a Vec3 the gradient is stored in Mat3, for e.x.:
-   *  Let \f$\boldsymbol{u} = \{u, v, w\} \f$ then,
-   *  \f$\left(\nabla \boldsymbol{u}\right)_i = \begin{bmatrix}\partial_x u & \partial_y u & \partial_z u \\ \partial_x
-   * v & \partial_y v & \partial_z v \\ \partial_x w & \partial_y w & \partial_z w \end{bmatrix}_i.\f$
-   *
-   * \param alpha Vec3 variable for which gradient will be computed.
-   * \param grad_alpha gradient of alpha is written to this.
-   */
-  void operator()(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha)
-  {
-    compute(alpha, grad_alpha);
-  }
+  static void compute(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha);
 };
 
 }  // namespace CPU
