+++ include/gals/cpu/gradient/third-order.h
@@ -31,8 +31,6 @@
 
 #pragma once
 
-#include "gals/cpu/gradient/second-order-central.h"
-#include "gals/cpu/gradient/third-order.h"
 #include "gals/utilities/array.h"
 #include "gals/utilities/grid.h"
 #include "gals/utilities/mat3.h"
@@ -42,23 +40,23 @@ namespace GALS
 {
 namespace CPU
 {
-/*! \class Gradient
+/*! \class ThirdOrder
  *
- * Class to create 3 component elements at a computational cell. For e.x. velocity, gradients, etc.
+ * Gradient computation using third order scheme.
  */
-template <typename T, typename T_GRID, typename GRADIENT_SCHEME = ThirdOrder<T, T_GRID>>
-class Gradient
+template <typename T, typename T_GRID>
+class ThirdOrder
 {
  public:
-  typedef T value_type;
+  using value_type = T;
 
   /*! Default constructor
    */
-  Gradient();
+  ThirdOrder();
 
   /*! Destructor
    */
-  ~Gradient();
+  ~ThirdOrder();
 
   /*! Compute gradient of scalar.
    *
@@ -68,19 +66,44 @@ class Gradient
    * \param alpha scalar variable for which gradient will be computed.
    * \param grad_alpha gradient of alpha is written to this.
    */
-  static void compute(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha);
+  void compute(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha);
 
-  /*! Compute gradient of vector.
+  /*! Overload operator to compute gradient of a scalar quantity.
+   *
+   * For a scalar the gradient is stored in Vec3, for e.x.:
+   *  \f$\left(\nabla \alpha\right)_i = \{\partial_x \alpha, \partial_y \alpha, \partial_z \alpha\}_i.\f$
+   *
+   * \param alpha scalar variable for which gradient will be computed.
+   * \param grad_alpha gradient of alpha is written to this.
+   */
+  void operator()(const Array<T_GRID, T> &alpha, Array<T_GRID, Vec3<T>> &grad_alpha) { compute(alpha, grad_alpha); }
+
+  /*! Compute gradient of Vec3.
    *
    * For a Vec3 the gradient is stored in Mat3, for e.x.:
    *  Let \f$\boldsymbol{u} = \{u, v, w\} \f$ then,
    *  \f$\left(\nabla \boldsymbol{u}\right)_i = \begin{bmatrix}\partial_x u & \partial_y u & \partial_z u \\ \partial_x
    * v & \partial_y v & \partial_z v \\ \partial_x w & \partial_y w & \partial_z w \end{bmatrix}_i.\f$
    *
-   * \param alpha vec3 variable for which gradient will be computed.
+   * \param alpha Vec3 variable for which gradient will be computed.
    * \param grad_alpha gradient of alpha is written to this.
    */
-  static void compute(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha);
+  void compute(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha);
+
+  /*! Overload operator to compute gradient of Vec3.
+   *
+   * For a Vec3 the gradient is stored in Mat3, for e.x.:
+   *  Let \f$\boldsymbol{u} = \{u, v, w\} \f$ then,
+   *  \f$\left(\nabla \boldsymbol{u}\right)_i = \begin{bmatrix}\partial_x u & \partial_y u & \partial_z u \\ \partial_x
+   * v & \partial_y v & \partial_z v \\ \partial_x w & \partial_y w & \partial_z w \end{bmatrix}_i.\f$
+   *
+   * \param alpha Vec3 variable for which gradient will be computed.
+   * \param grad_alpha gradient of alpha is written to this.
+   */
+  void operator()(const Array<T_GRID, Vec3<T>> &alpha, Array<T_GRID, Mat3<T>> &grad_alpha)
+  {
+    compute(alpha, grad_alpha);
+  }
 };
 
 }  // namespace CPU
