+++ include/gals/utilities/grid.h
@@ -35,13 +35,31 @@
 #include <vector>
 
 #include "gals/input-fields/grid.h"
+#include "gals/utilities/utilities.h"
 #include "mat3.h"
-#include "utilities/vec3.h"
+#include "vec3.h"
 
 namespace GALS
 {
 namespace CPU
 {
+constexpr int num_mixed_derivatives(const int dim)
+{
+  switch (dim) {
+    case 1:
+      return 0;
+      break;
+    case 2:
+      return 1;
+      break;
+    case 3:
+      return 4;
+      break;
+    default:
+      return -1;
+  }
+}
+
 /*! \class Grid
  *
  * Class to create grid.
@@ -50,12 +68,17 @@ template <typename T, int DIM = 3>
 class Grid
 {
  public:
-  typedef T value_type;
-  typedef Vec3<T> position_type;
-  static const int dim = DIM;
+  using value_type = T;
+  using position_type = Vec3<T>;
+
+  //! Dimension.
+  static constexpr int dim = DIM;
 
   //! vector along x: {1, 0, 0}; y: {0, 1, 0}; z: {0, 0, 1}.
   static const Mat3<int> axis_vectors;
+
+  //! Number of mixed derivatives.
+  static constexpr int num_mixed_derivatives = GALS::CPU::num_mixed_derivatives(DIM);
 
   /*! Constructor with arguments for number of cells across x, y, z.
    *
