+++ tests/CMakeLists.txt
@@ -49,7 +49,7 @@ SET_PROPERTY(TARGET ${TEST_NAME} PROPERTY CXX_STANDARD 14)
 TARGET_INCLUDE_DIRECTORIES(${TEST_NAME}
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
   ${PROJECT_SOURCE_DIR}/src
-  ${PROJECT_SOURCE_DIR}/src/include
+  ${PROJECT_SOURCE_DIR}/include
 )
 
 TARGET_LINK_LIBRARIES(${TEST_NAME}
